fix: comment fixes: dependency fixes for react-redux. Check StateManager initialization in extension. Ternary op fix. Scss import fix. Moved extension to a separate file.
