+++ packages/property-grid/src/PropertyGridManager.ts
@@ -5,14 +5,9 @@
 
 import * as i18next from "i18next";
 
-import { I18N, I18NNamespace } from "@bentley/imodeljs-i18n";
-import {
-  getClassName,
-  UiError,
-  UiItemsManager,
-} from "@bentley/ui-abstract";
-import { Extension, IModelApp } from "@bentley/imodeljs-frontend";
-import { PropertyGridUiItemsProvider } from "./components/PropertyGridUiItemsProvider";
+import { I18N } from "@bentley/imodeljs-i18n";
+import { getClassName, UiError } from "@bentley/ui-abstract";
+import { StateManager } from "@bentley/ui-framework";
 
 /**
  * Entry point for static initialization required by various components used in the package.
@@ -26,6 +21,12 @@ export class PropertyGridManager {
    * @param i18n - The internationalization service created by the IModelApp.
    */
   public static async initialize(i18n: I18N): Promise<void> {
+    if (!StateManager.isInitialized()) {
+      throw new Error(
+        "UiFramework's StateManager must be initialized for Property Grid to work properly as an extension",
+      );
+    }
+
     PropertyGridManager._i18n = i18n;
     return PropertyGridManager._i18n.registerNamespace(
       PropertyGridManager.i18nNamespace,
@@ -79,33 +80,4 @@ export class PropertyGridManager {
       PropertyGridManager.packageName + (className ? `.${className}` : "");
     return category;
   }
-}
-
-/** Extension object for loading on runtime */
-class PropertyGridExtension extends Extension {
-  protected _defaultNs = "PropertyGrid";
-  private _i18NNamespace?: I18NNamespace;
-
-  public async onExecute(_args: string[]): Promise<void> {
-    // No-op
-  }
-
-  public async onLoad(_args: string[]): Promise<void> {
-    // TODO:
-    // Register namespace
-    this._i18NNamespace = this.i18n.getNamespace(
-      PropertyGridManager.i18nNamespace,
-    );
-    if (this._i18NNamespace === undefined) {
-      throw new Error("Property grid extension could not find locale");
-    }
-    await this._i18NNamespace.readFinished;
-    await PropertyGridManager.initialize(this.i18n);
-    UiItemsManager.register(new PropertyGridUiItemsProvider());
-    // Register item provider
-  }
-}
-
-if (IModelApp.extensionAdmin) {
-  IModelApp.extensionAdmin.register(new PropertyGridExtension("propertyGrid"));
 }
