+++ packages/property-grid/src/components/ElementList.scss
@@ -2,9 +2,7 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
-@import "~@bentley/ui-core/lib/ui-core/style/index";
-@import "~@bentley/ui-core/lib/ui-core/text/index";
-@import "@bentley/ui-core/lib/ui-core/index";
+@import "~@bentley/ui-core/lib/ui-core/index";
 
 .property-grid-react-element-list {
   height: 100%;
