+++ packages/property-grid/src/property-grid-react.ts
@@ -3,5 +3,6 @@
 * Licensed under the MIT License. See LICENSE.md in the project root for license terms.
 *--------------------------------------------------------------------------------------------*/
 export * from "./PropertyGridManager";
+export * from "./PropertyGridExtension";
 export * from "./components/PropertyGrid";
 export * from "./components/MultiElementPropertyGrid";
